+++ crates/objc2/src/ffi/constants.rs
@@ -4,7 +4,8 @@
 #[cfg(any(doc, target_vendor = "apple"))]
 use std::os::raw::c_int;
 
-use crate::ffi::{id, objc_class, BOOL};
+use crate::ffi::{id, BOOL};
+use crate::runtime::AnyClass;
 
 /// The equivalent of `true` for Objective-C's [`BOOL`][`super::BOOL`] type.
 #[allow(clippy::unnecessary_cast)]
@@ -18,7 +19,7 @@ pub const NO: BOOL = false as BOOL; // false -> 0
 pub const nil: id = 0 as *mut _;
 
 /// A quick alias for a [`null_mut`][`core::ptr::null_mut`] class.
-pub const Nil: *mut objc_class = 0 as *mut _;
+pub const Nil: *mut AnyClass = 0 as *mut _;
 
 /// Policies related to associative references.
 ///