+++ crates/objc2/src/ffi/various.rs
@@ -29,12 +29,12 @@ pub type IMP = Option<InnerImp>;
 // /// Remember that this is non-null!
 // #[cfg(any(doc, all(target_vendor = "apple", not(all(target_os = "macos", target_arch = "x86")))))]
 // pub type objc_hook_getClass =
-//     unsafe extern "C" fn(name: *const c_char, out_cls: *mut *const crate::ffi::objc_class) -> BOOL;
+//     unsafe extern "C" fn(name: *const c_char, out_cls: *mut *const crate::runtime::AnyClass) -> BOOL;
 //
 // /// Remember that this is non-null!
 // #[cfg(any(doc, all(target_vendor = "apple", not(all(target_os = "macos", target_arch = "x86")))))]
 // pub type objc_hook_lazyClassNamer =
-//     unsafe extern "C" fn(cls: *const crate::ffi::objc_class) -> *const c_char;
+//     unsafe extern "C" fn(cls: *const crate::runtime::AnyClass) -> *const c_char;
 
 extern_c_unwind! {
     // Instead of being able to change this, it's a weak symbol on GNUStep.